Implement bedrock raw chunk API (#21)
* Add BedrockRawChunk class

* Redesign chunk API

* Redesign chunk API

* Add include

* Add missing pragma once

* Implement BedrockRawDimension::has_chunk

* Bedrock: Refactor chunk and add get_raw_chunk

* Implement delete chunk

* Add delete chunk bindings

* Add for_keys_in_chunk

This code needs to be used a couple of times so make it reusable.

* Add a lock for the Java _level_dat attribute

* Add a lock to Bedrock _level_dat

* Add const member getters

* Convert KeyWriter to a template

* Move methods into cpp file

* Move the locks into a pointer

Some functions require locking the mutex but are otherwise const methods.
The only solution I can see is to move the mutex into a pointer (out of the class)

* Implement chunk writing

* Remove redundant increment

* Update stubs

* Add get, set, encode and decode chunk bindings

* Simplify module constructors

* Expand LevelLoaderPathToken constructor

* Implement Bedrock chunk coord iterator

* Remove Java level_dat lock

Upon further inspection this is handled by the public lock

* Remove Bedrock level_dat lock

* Bump dependencies

+++ src/amulet/level/bedrock/chunk_encode.cpp
@@ -10,8 +10,8 @@ namespace Amulet {
 
 BedrockRawChunk BedrockRawDimension::encode_chunk(
     BedrockChunk& chunk,
-    std::int64_t cx,
-    std::int64_t cz)
+    std::int32_t cx,
+    std::int32_t cz)
 {
     throw std::runtime_error("NotImplementedError");
 }
